#96 音SE(仮)のエフェクト種類を3つまで増やす。
