+++ PicGather/Assets/Effect/Sound/SoundEffectController.cs.meta
